Update composer to 1.1.0-RC stable
This will fix the issue legoktm and hashar had because composer was not
compatible at the time with symfony 3.x.

This fixes issues with symfony 2.x and 3.x.

We are going to use symfony 2.x to stay compatible with php 5.3 or else
composer-php53 will fail. and other composer related tools for php 5.3
since 3.x is not compatible with php 5.3.

Composer was promoted to stable now.

Changelog is at
https://github.com/composer/composer/blob/1.1.0-RC/CHANGELOG.md

Bug: T125343
Change-Id: Ic71fd58bc0cc94f3a76937c349315067b95be378