Update composer to 1.0.3 stable
This will fix the issue legoktm and hashar had because composer was not
compatible at the time with symfony 3.x.

This fixes issues with symfony 2.x and 3.x.

We are going to use symfony 2.x to stay compatible with php 5.3 or else
composer-php53 will fail. and other composer related tools for php 5.3
since 3.x is not compatible with php 5.3.

Composer was promoted to stable now.

Changelog is at
https://github.com/composer/composer/blob/1.0.3/CHANGELOG.md

Using the previous composer version after the next one in
8a176a28da0d967fdfc2f3426428941b59db6d68 produced an autoloader that was not
PHP 5.5 compatible.

Bug: T125343
Change-Id: I81eb0ed3bcf8efdd60ab24a39d435a89647da35e